Helidon 2.0 and cleaner separation of GCS specific integration
