+++ src/main/java/no/ssb/dapla/metadata/distributor/MetadataDistributorApplicationBuilder.java
@@ -1,7 +1,10 @@
 package no.ssb.dapla.metadata.distributor;
 
+import io.grpc.LoadBalancerRegistry;
+import io.grpc.NameResolverRegistry;
+import io.grpc.internal.DnsNameResolverProvider;
+import io.grpc.internal.PickFirstLoadBalancerProvider;
 import io.helidon.config.Config;
-import io.helidon.config.ConfigSources;
 import io.helidon.config.spi.ConfigSource;
 import io.helidon.tracing.TracerBuilder;
 import io.opentracing.Tracer;
@@ -10,24 +13,40 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.function.Supplier;
 
+import static io.helidon.config.ConfigSources.classpath;
+import static io.helidon.config.ConfigSources.file;
 import static java.util.Optional.ofNullable;
 
 public class MetadataDistributorApplicationBuilder {
     protected Config config;
 
+    public static void applyGrpcProvidersWorkaround() {
+        // The shaded version of grpc from helidon does not include the service definition for
+        // PickFirstLoadBalancerProvider. This result in LoadBalancerRegistry not being able to
+        // find it. We register them manually here.
+        LoadBalancerRegistry.getDefaultRegistry().register(new PickFirstLoadBalancerProvider());
+        // LoadBalancerRegistry.getDefaultRegistry().register(new HealthCheckingRoundRobinLoadBalancerProvider());
+
+        // The same thing happens with the name resolvers.
+        NameResolverRegistry.getDefaultRegistry().register(new DnsNameResolverProvider());
+    }
+
     public static Config createDefaultConfig() {
+        Config.Builder builder = Config.builder();
         List<Supplier<ConfigSource>> configSourceSupplierList = new LinkedList();
         String overrideFile = System.getenv("HELIDON_CONFIG_FILE");
         if (overrideFile != null) {
-            configSourceSupplierList.add(ConfigSources.file(overrideFile).optional());
+            builder.addSource(file(overrideFile).optional());
         }
-
-        configSourceSupplierList.add(ConfigSources.file("conf/application.yaml").optional());
-        configSourceSupplierList.add(ConfigSources.classpath("application.yaml"));
-        return Config.builder().sources(configSourceSupplierList).build();
+        return builder
+                .addSource(file("conf/application.yaml").optional())
+                .addSource(classpath("application.yaml"))
+                .build();
     }
 
     public MetadataDistributorApplication build() {
+        applyGrpcProvidersWorkaround();
+
         Config config = ofNullable(this.config).orElseGet(() -> createDefaultConfig());
 
         TracerBuilder<?> tracerBuilder = TracerBuilder.create(config.get("tracing")).registerGlobal(false);
