+++ src/main/java/no/ssb/dapla/metadata/distributor/dataset/MetadataRouter.java
@@ -7,65 +7,50 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutureCallback;
 import com.google.api.core.ApiFutures;
-import com.google.api.gax.paging.Page;
 import com.google.cloud.pubsub.v1.AckReplyConsumer;
 import com.google.cloud.pubsub.v1.MessageReceiver;
 import com.google.cloud.pubsub.v1.Publisher;
 import com.google.cloud.pubsub.v1.Subscriber;
-import com.google.cloud.storage.Blob;
-import com.google.cloud.storage.BlobId;
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.Storage.BlobListOption;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.google.protobuf.ByteString;
 import com.google.pubsub.v1.PubsubMessage;
 import io.helidon.config.Config;
 import no.ssb.dapla.dataset.api.DatasetMeta;
 import no.ssb.dapla.dataset.uri.DatasetUri;
-import no.ssb.dapla.metadata.distributor.parquet.GCSReadChannelBasedInputFile;
-import no.ssb.dapla.metadata.distributor.parquet.ParquetTools;
 import no.ssb.helidon.media.protobuf.ProtobufJsonUtils;
 import no.ssb.pubsub.PubSub;
-import org.apache.avro.Schema;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.parquet.hadoop.util.HadoopInputFile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Predicate;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.stream.StreamSupport;
-
-import static java.util.Optional.ofNullable;
 
 public class MetadataRouter {
 
     private static final Logger LOG = LoggerFactory.getLogger(MetadataRouter.class);
 
     final PubSub pubSub;
-    final Storage storage;
+    final Map<String, DatasetStore> datasetStoreByScheme = new ConcurrentHashMap<>();
     final List<Subscriber> subscribers = new CopyOnWriteArrayList<>();
     final List<Publisher> publishers = new CopyOnWriteArrayList<>();
     final MetadataSignatureVerifier metadataSignatureVerifier;
-    final String fileSystemDataFolder;
 
-    public MetadataRouter(Config routeConfig, PubSub pubSub, Storage storage, MetadataSignatureVerifier metadataSignatureVerifier, String fileSystemDataFolder) {
+    public MetadataRouter(Config routeConfig, PubSub pubSub, MetadataSignatureVerifier metadataSignatureVerifier, DatasetStore... datasetStores) {
         this.pubSub = pubSub;
-        this.storage = storage;
         this.metadataSignatureVerifier = metadataSignatureVerifier;
-        this.fileSystemDataFolder = fileSystemDataFolder;
+
+        for (DatasetStore datasetStore : datasetStores) {
+            datasetStoreByScheme.put(datasetStore.supportedScheme(), datasetStore);
+        }
 
         List<Config> upstreams = routeConfig.get("upstream").asNodeList().get();
         List<Config> downstreams = routeConfig.get("downstream").asNodeList().get();
@@ -104,132 +89,6 @@ public class MetadataRouter {
         }
     }
 
-    static MetadataReadAndVerifyResult resolveAndReadDatasetMeta(
-            Storage storage,
-            String fileSystemDataFolder,
-            MetadataSignatureVerifier metadataSignatureVerifier,
-            DatasetUri datasetUri
-    ) throws IOException {
-        byte[] datasetMetaBytes;
-        byte[] datasetDocBytes = null;
-        byte[] datasetLineageBytes = null;
-        byte[] datasetMetaSignatureBytes;
-        String datasetMetaJsonPath = datasetUri.toURI().getPath() + "/.dataset-meta.json";
-        String datasetDocJsonPath = datasetUri.toURI().getPath() + "/.dataset-doc.json";
-        String datasetLinageJsonPath = datasetUri.toURI().getPath() + "/.dataset-lineage.json";
-        String datasetMetaJsonSignaturePath = datasetUri.toURI().getPath() + "/.dataset-meta.json.sign";
-        String scheme = datasetUri.toURI().getScheme();
-        Schema avroSchema;
-        switch (scheme) {
-            case "file":
-                datasetMetaBytes = Files.readAllBytes(Path.of(fileSystemDataFolder, datasetMetaJsonPath));
-                if (Files.isReadable(Path.of(fileSystemDataFolder, datasetDocJsonPath))) {
-                    datasetDocBytes = Files.readAllBytes(Path.of(fileSystemDataFolder, datasetDocJsonPath));
-                }
-                if (Files.isReadable(Path.of(fileSystemDataFolder, datasetLinageJsonPath))) {
-                    datasetLineageBytes = Files.readAllBytes(Path.of(fileSystemDataFolder, datasetLinageJsonPath));
-                }
-                datasetMetaSignatureBytes = Files.readAllBytes(Path.of(fileSystemDataFolder, datasetMetaJsonSignaturePath));
-                avroSchema = getAvroSchemaFromLocalFileSystem(datasetUri);
-                break;
-            case "gs":
-                String bucket = datasetUri.toURI().getHost();
-                datasetMetaBytes = storage.readAllBytes(BlobId.of(bucket, stripLeadingSlashes(datasetMetaJsonPath)));
-                BlobId datasetDocBlobId = BlobId.of(bucket, stripLeadingSlashes(datasetDocJsonPath));
-                Blob datasetDocBlob = storage.get(datasetDocBlobId);
-                if (datasetDocBlob != null) {
-                    datasetDocBytes = storage.readAllBytes(datasetDocBlobId);
-                }
-                BlobId datasetLineageBlobId = BlobId.of(bucket, stripLeadingSlashes(datasetLinageJsonPath));
-                Blob datasetLineageBlob = storage.get(datasetLineageBlobId);
-                if (datasetLineageBlob != null) {
-                    datasetLineageBytes = storage.readAllBytes(datasetLineageBlobId);
-                }
-
-                datasetMetaSignatureBytes = storage.readAllBytes(BlobId.of(bucket, stripLeadingSlashes(datasetMetaJsonSignaturePath)));
-                avroSchema = getAvroSchemaFromGoogleCloudStorage(storage, datasetUri);
-                break;
-            default:
-                throw new RuntimeException("Scheme not supported. scheme='" + scheme + "'");
-        }
-        boolean verified = metadataSignatureVerifier.verify(datasetMetaBytes, datasetMetaSignatureBytes);
-        return new MetadataReadAndVerifyResult(verified,
-                ByteString.copyFrom(datasetMetaBytes),
-                ofNullable(datasetDocBytes).map(ByteString::copyFrom).orElse(null),
-                ofNullable(datasetLineageBytes).map(ByteString::copyFrom).orElse(null),
-                avroSchema
-        );
-    }
-
-    private static Schema getAvroSchemaFromLocalFileSystem(DatasetUri datasetUri) {
-        org.apache.hadoop.fs.Path hadoopPath = new org.apache.hadoop.fs.Path(datasetUri.toURI().getRawPath());
-        HadoopInputFile hadoopInputFile;
-        try {
-            hadoopInputFile = HadoopInputFile.fromPath(hadoopPath, new Configuration());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        Schema schema = ParquetTools.getAvroSchemaFromFile(hadoopInputFile);
-        return schema;
-    }
-
-    private static Schema getAvroSchemaFromGoogleCloudStorage(Storage storage, DatasetUri datasetUri) {
-        String bucket = datasetUri.toURI().getHost();
-        String prefix = datasetUri.toURI().getPath();
-        Page<Blob> page = storage.list(bucket, BlobListOption.prefix(prefix), BlobListOption.pageSize(10));
-        Blob firstParquetBlob = paginateUntil(page, b -> b.getName().endsWith(".parquet"));
-        if (firstParquetBlob == null) {
-            return null;
-        }
-        Schema schema = ParquetTools.getAvroSchemaFromFile(new GCSReadChannelBasedInputFile(firstParquetBlob));
-        return schema;
-    }
-
-    private static Blob paginateUntil(Page<Blob> firstPage, Predicate<? super Blob> predicate) {
-        return StreamSupport.stream(firstPage.iterateAll().spliterator(), false)
-                .filter(predicate)
-                .findFirst()
-                .orElseGet(() -> {
-                    Page<Blob> page = firstPage;
-                    while (page.hasNextPage()) {
-                        page = page.getNextPage();
-                        Optional<Blob> first = StreamSupport.stream(page.iterateAll().spliterator(), false)
-                                .filter(predicate)
-                                .findFirst();
-                        if (first.isPresent()) {
-                            return first.get();
-                        }
-                    }
-                    return null;
-                });
-    }
-
-    private static String stripLeadingSlashes(String input) {
-        return input.startsWith("/") ? stripLeadingSlashes(input.substring(1)) : input;
-    }
-
-    static class MetadataReadAndVerifyResult {
-
-        final boolean signatureValid;
-        final ByteString datasetMetaByteString;
-        final ByteString datasetDocByteString;
-        final ByteString datasetLineageByteString;
-        final Schema schema;
-
-        MetadataReadAndVerifyResult(boolean signatureValid,
-                                    ByteString datasetMetaByteString,
-                                    ByteString datasetDocByteString,
-                                    ByteString datasetLineageByteString,
-                                    Schema schema) {
-            this.signatureValid = signatureValid;
-            this.datasetMetaByteString = datasetMetaByteString;
-            this.datasetDocByteString = datasetDocByteString;
-            this.datasetLineageByteString = datasetLineageByteString;
-            this.schema = schema;
-        }
-
-    }
-
     static final ObjectMapper objectMapper = new ObjectMapper();
 
     class RouterMessageReceiver implements MessageReceiver {
@@ -246,14 +105,12 @@ public class MetadataRouter {
 
         @Override
         public void receiveMessage(PubsubMessage upstreamMessage, AckReplyConsumer consumer) {
-            process(storage, fileSystemDataFolder, metadataSignatureVerifier, publishers, topic, subscription, upstreamMessage, consumer::ack);
+            process(datasetStoreByScheme, publishers, topic, subscription, upstreamMessage, consumer::ack);
         }
     }
 
     static void process(
-            Storage storage,
-            String fileSystemDataFolder,
-            MetadataSignatureVerifier metadataSignatureVerifier,
+            Map<String, DatasetStore> datasetStoreByScheme,
             List<Publisher> publishers,
             String topic,
             String subscription,
@@ -319,7 +176,13 @@ public class MetadataRouter {
 
             DatasetUri datasetUri = DatasetUri.of(schemeAndAuthority, path, version);
 
-            MetadataReadAndVerifyResult metadataReadAndVerifyResult = resolveAndReadDatasetMeta(storage, fileSystemDataFolder, metadataSignatureVerifier, datasetUri);
+            String scheme = datasetUri.toURI().getScheme();
+            DatasetStore datasetStore = datasetStoreByScheme.get(scheme);
+            if (datasetStore == null) {
+                throw new IllegalArgumentException("Unsupported scheme: " + scheme);
+            }
+            MetadataReadAndVerifyResult metadataReadAndVerifyResult = datasetStore.resolveAndReadDatasetMeta(datasetUri);
+
             if (!metadataReadAndVerifyResult.signatureValid) {
                 LOG.warn("Invalid signature for metadata of dataset: {}", datasetUri.toString());
                 ack.run();
