+++ src/main/java/no/ssb/helidon/application/Tracing.java
@@ -1,16 +1,15 @@
 package no.ssb.helidon.application;
 
 import com.google.protobuf.MessageOrBuilder;
-import io.helidon.common.context.Contexts;
 import io.helidon.webserver.ServerRequest;
 import io.opentracing.Span;
-import io.opentracing.SpanContext;
 import io.opentracing.Tracer;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Optional;
 
 import static java.util.Optional.ofNullable;
 
@@ -67,39 +66,14 @@ public class Tracing {
         tracer.scopeManager().activate(span);
     }
 
-    public static void restoreTracingContext(TracerAndSpan tracerAndSpan) {
-        restoreTracingContext(tracerAndSpan.tracer, tracerAndSpan.span);
-    }
-
-    public static <T extends MessageOrBuilder> TracerAndSpan spanFromGrpc(T message, String operationName) {
-        Tracer tracer = Contexts
-                .context()
-                .get()
-                .get(Tracer.class)
-                .get();
-        if (tracer == null) {
-            throw new IllegalStateException("A Tracer has not been assigned to the Helidon Contexts context");
-        }
-        SpanContext spanContext = Contexts.context()
-                .get()
-                .get(SpanContext.class)
-                .get();
-        Span span = tracer
-                .buildSpan(operationName)
-                .asChildOf(spanContext)
-                .start();
-        tracer.scopeManager().activate(span);
-        return new TracerAndSpan(tracer, span);
-    }
-
-    public static TracerAndSpan spanFromHttp(ServerRequest request, String operationName) {
-        SpanContext spanContext = request.spanContext();
-        Tracer tracer = request.tracer();
-        Span span = tracer
-                .buildSpan(operationName)
-                .asChildOf(spanContext)
-                .start();
-        tracer.scopeManager().activate(span);
-        return new TracerAndSpan(tracer, span);
+    public static Optional<Span> spanFromHttp(ServerRequest request, String operationName) {
+        return request.spanContext()
+                .map(ctx -> request.tracer()
+                        .buildSpan(operationName)
+                        .asChildOf(ctx)
+                        .start()
+                )
+                .map(Optional::of)
+                .orElse(Optional.empty());
     }
 }
